Add hideCloseButton with false default
