+++ client/src/app/pages/applications/application-assessment/application-assessment.tsx
@@ -210,7 +210,6 @@ export const ApplicationAssessment: React.FC = () => {
                 pushNotification({
                   title: getAxiosErrorMessage(error),
                   variant: "danger",
-                  actionClose: true,
                 });
                 formikHelpers.setSubmitting(false);
               });
