+++ client/src/app/pages/applications/application-review/application-review.tsx
@@ -125,7 +125,6 @@ export const ApplicationReview: React.FC = () => {
         pushNotification({
           title: getAxiosErrorMessage(error),
           variant: "danger",
-          actionClose: true,
         });
       }
     );
