+++ client/src/app/shared/components/Notifications.tsx
@@ -6,22 +6,6 @@ import {
 } from "@patternfly/react-core";
 import { NotificationsContext } from "../notifications-context";
 
-export type notificationType =
-  | "success"
-  | "info"
-  | "warning"
-  | "danger"
-  | "default";
-
-export interface INotification {
-  title: string;
-  variant: notificationType;
-  key?: string;
-  message?: string | JSX.Element;
-  actionClose?: boolean;
-  timeout?: number | boolean;
-}
-
 export const Notifications: React.FunctionComponent = () => {
   const appContext = React.useContext(NotificationsContext);
   return (
@@ -35,7 +19,7 @@ export const Notifications: React.FunctionComponent = () => {
             title={notification.title}
             variant={notification.variant}
             key={key}
-            {...(notification.actionClose && {
+            {...(!notification.hideCloseButton && {
               actionClose: (
                 <AlertActionCloseButton
                   onClose={() => {
