+++ client/src/app/shared/containers/bulk-copy-notifications-container/bulk-copy-notifications-container.tsx
@@ -103,7 +103,6 @@ export const BulkCopyNotificationsContainer: React.FC = () => {
           ? t("toastr.success.assessmentAndReviewCopied")
           : t("toastr.success.assessmentCopied"),
         variant: "success",
-        actionClose: true,
       });
 
       queryClient.invalidateQueries(assessmentsQueryKey);
